% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_all_tyfs.R
\name{build_all_ywccms}
\alias{build_all_ywccms}
\title{Build Yearly Weighted Condition Contribution Matrices (YWCCMs)}
\usage{
build_all_ywccms(raw_cis, year, year_list, ciwms_list)
}
\arguments{
\item{raw_cis}{A data frame of raw condition indicator scores.}

\item{year}{The year for which to build the matrices.}

\item{year_list}{A vector of all years in the account (to establish base year).}

\item{ciwms_list}{A named list of Condition Indicator Weighting Matrices (CIWMs).}
}
\value{
A named list of matrices (YWCCMs), one for each indicator.
}
\description{
Generates a list of matrices where each indicator's relevance weighting
(CIWM) is multiplied by its indexed condition score for a given year.
}
\keyword{internal}
