% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ciwm_list.R
\name{build_ciwm_list}
\alias{build_ciwm_list}
\title{Build Condition Indicator Weight Matrices (CIWM)}
\usage{
build_ciwm_list(
  cirm_list,
  indicator_directory,
  es_label_tree,
  habitats_label_tree = NULL
)
}
\arguments{
\item{cirm_list}{A named list of data frames. Names are condition indicator
labels. Data frames hold a matrix of binary values denoting whether the
Condition Indicator is relevant to each habitat/service combination. Names
must match the labels in \code{indicator_directory}.}

\item{indicator_directory}{A data frame containing weights
denoting the relevance of each condition indicator to each ecosystem service
type. Must include a column of condition indicator labels for matching.}

\item{es_label_tree}{A named list of character vectors representing
the hierarchy of ecosystem services (as character vectors) within
service type group (as list object names).
Syntactical names only (no spaces or special characters).}

\item{habitats_label_tree}{A named list of character vectors representing the
hierarchy of habitats (as character vectors, typically
EUNIS Level 2) within broad habitats (as list object names, typically EUNIS
Level 1). Syntactical names only (no spaces or special characters).
The habitats label tree is optional; if supplied a labelled data frame will
be returned.}
}
\value{
A named list of data frames, one list object for each Condition
Indicator each containing the weighted relevance of that indicator to each
combination of habitats and ecosystem services.
}
\description{
Converts relevance matrices (CIRM) into weight matrices (CIWM) by
multiplying the relevance scores by the indicator's importance weights
(per ecosystem service type) as defined in the Indicator Directory.
}
