% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_es_potential_base.R
\name{calc_es_potential_base}
\alias{calc_es_potential_base}
\title{Calculate the Ecosystem Service Potential Base}
\usage{
calc_es_potential_base(
  provision_per_unit_weights,
  habitat_extent,
  year_list,
  year_one = NULL,
  habitats_label_tree,
  es_label_tree
)
}
\arguments{
\item{provision_per_unit_weights}{A data frame containing the ecosystem service provision
potential per unit weights. Row order must match the expanded habitats label
tree and column order must match the expanded ES label tree.}

\item{habitat_extent}{A data frame containing data representing the extent
(area) of each habitat in each year of the index. Rows = habitats. Columns =
years. There must be a column to match the identified baseline year. Row
order must match the expanded habitats label tree.}

\item{year_list}{A vector (character or numeric) of the years over which the
index is to be calculated.}

\item{year_one}{Optional. The specific year from \code{habitat_extent} to
use as the baseline for extent. If \code{NULL} (default), uses the first
element of \code{year_list}.}

\item{habitats_label_tree}{A named list of character vectors representing the
hierarchy of habitats.}

\item{es_label_tree}{A named list of character vectors representing
the hierarchy of ecosystem services.}
}
\value{
A labelled data frame with the same dimensions as 'provision_per_unit_weights'.
}
\description{
Multiply the Provision Per Unit (weights) by the
habitat extent in the baseline year of the index.
}
\keyword{internal}
