% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_espb.R
\name{calc_espb}
\alias{calc_espb}
\title{Calculate the Ecosystem Service Potential Base (ESPB)}
\usage{
calc_espb(
  esppu_weights,
  habitat_extent,
  year_list,
  year_one = NULL,
  habitats_label_tree,
  es_label_tree
)
}
\arguments{
\item{esppu_weights}{A data frame containing the ecosystem service provision
potential per unit weights. Row order must match the expanded habitats label
tree and column order must match the expanded ES label tree.}

\item{habitat_extent}{A data frame containing data representing the extent
(area) of each habitat in each year of the index. Rows = habitats. Columns =
years. There must be a column to match the identified baseline year. Row
order must match the expanded habitats label tree.}

\item{year_list}{A vector (character or numeric) of the years over which the
index is to be calculated.}

\item{year_one}{Optional. The specific year from \code{habitat_extent} to
use as the baseline for extent. If \code{NULL} (default), uses the first
element of \code{year_list}.}

\item{habitats_label_tree}{A named list of character vectors representing the
hierarchy of habitats.}

\item{es_label_tree}{A named list of character vectors representing
the hierarchy of ecosystem services.}
}
\value{
A labelled data frame with the same dimensions as 'esppu_weights'.
}
\description{
Multiply the ecosystem service provision potential per unit (ESPPU) by the
habitat extent in the baseline year of the index.
}
\examples{
# Define Label Trees
h_tree <- list(coastal = c("b1", "b2"), woodland = c("g1"))
es_tree <- list(provisioning = c("crops", "timber"))

# Setup Habitat Extent
extent <- data.frame(
  `2026` = c(100, 150, 200),
  `2027` = c(110, 140, 210),
  check.names = FALSE
)

# Setup ESPPU Weights
weights <- data.frame(crops = c(0.12, 0.1, 0.0), timber = c(0.0, 0.0, 0.9))

years <- c("2026", "2027")

# Run with default (2026)
espb_res <- calc_espb(weights, extent, years, habitats_label_tree = h_tree, es_label_tree = es_tree)
}
