% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncai.R
\name{calc_ncai_by_bh}
\alias{calc_ncai_by_bh}
\title{Calculate NCAI Broken Down by Broad Habitat}
\usage{
calc_ncai_by_bh(
  total_assets_matrix_list,
  habitats_label_tree,
  year_one = names(total_assets_matrix_list)[[1]],
  ...
)
}
\arguments{
\item{total_assets_matrix_list}{A named list of annual asset data frames.}

\item{habitats_label_tree}{A named list where each element is a character
vector of habitat labels (row names).}

\item{year_one}{Optional: year to index around. Default is year one of the
\code{year_list}.}

\item{...}{Additional arguments passed to \code{calc_ncai}.}
}
\value{
A list of NCAI data frames, one for each habitat group.
}
\description{
Calculates the index subsetted by broad habitat
}
\keyword{internal}
