% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_wellbeing_potential_base.R
\name{calc_wellbeing_potential_base}
\alias{calc_wellbeing_potential_base}
\title{Scales the Ecosystem Service (ES) Potential Base by the Importance Weight
for each service to generate the Wellbeing Base. ES Potential Base is expressed as each
habitat's contribution as a proportion of the total potential for that
service. Label trees may be passed in for optional labelling of the output
data frame.}
\usage{
calc_wellbeing_potential_base(
  es_potential_base,
  importance_weights,
  habitats_label_tree = NULL,
  es_label_tree = NULL
)
}
\arguments{
\item{es_potential_base}{A matrix (or data frame) of habitat (rows) by ecosystem service
(columns) containing weights to denote the potential of each habitat to
provide each ecosystem service, calculated for the baseline year (Year One).}

\item{importance_weights}{A vector of weights denoting the importance of
each ecosystem service in the area of interest.}

\item{habitats_label_tree}{A named list of character vectors representing the
hierarchy of habitats (as character vectors, typically
EUNIS Level 2) within broad habitats (as list object names, typically EUNIS
Level 1). Syntactical names only (no spaces or special characters).
The habitats label tree is optional and if supplied will be used to label
the returned data frame.}

\item{es_label_tree}{A named list of character vectors representing
the hierarchy of ecosystem services (as character vectors) within
service type group (as list object names).
Syntactical names only (no spaces or special characters).
The ES label tree is optional and if supplied will be used to label the
returned data frame.}
}
\value{
A data frame representing the Wellbeing Base.
}
\description{
Scales the Ecosystem Service (ES) Potential Base by the Importance Weight
for each service to generate the Wellbeing Base. ES Potential Base is expressed as each
habitat's contribution as a proportion of the total potential for that
service. Label trees may be passed in for optional labelling of the output
data frame.
}
\examples{
# Assuming es_potential_base and importance_weights are already calculated:
# wellbeing_potential_base_matrix <- calc_wellbeing_potential_base(
#   es_potential_base = es_potential_base_matrix,
#   importance_weights = weights_vector,
#   habitats_label_tree = habitats_tree,
#   es_label_tree = es_tree
# )
}
\keyword{internal}
