% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esppu_scores_to_weights.R
\name{esppu_scores_to_weights}
\alias{esppu_scores_to_weights}
\title{Transform ESPPU Scores to Weights}
\usage{
esppu_scores_to_weights(
  esppu,
  divisor = NULL,
  custom_divisor_matrix = NULL,
  habitats_label_tree,
  es_label_tree
)
}
\arguments{
\item{esppu}{A numeric \code{data.frame} where rows represent habitat types
and columns represent ecosystem services and cells denote the potential of
the relevant habitat type to provide the relevant ecosystem service.}

\item{divisor}{A single numeric value by which ES potential scores are
divided to calculate the weights. Only used if \code{custom_divisor_matrix}
is \code{NULL}.}

\item{custom_divisor_matrix}{A numeric \code{data.frame} of the same
dimensions as \code{esppu}, containing the number by which the score should
be divided to generate the weight for each habitat/service
combination.}

\item{habitats_label_tree}{A named list of character vectors where
each list name represents a broad habitat category, typically a EUNIS level
1 habitat (e.g. "coastal") and the associated character vector contains the
labels or codes of the habitat sub-types - typically EUNIS level two
habitats - falling within that category (e.g., c("b1", "b2") or
c("coastal_dunes_sandy_shores", "coastal_shingle")). Syntactical names only
(no spaces or special characters).
The habitats label tree defines the relevant habitats for calculating the
NCAI and will be used to label the returned data frame.}

\item{es_label_tree}{A named list of character vectors where each list
name represents a label name for a type group of ecosystem services
(e.g. 'provisioning') and the associated character vector contains the
labels for the ecosystem services in that group (e.g. 'cultivated_crops').
Syntactical names only (no spaces or special characters).
The ES label tree defines the relevant ecosystem services for calculating
the NCAI and will be used to label the returned data frame.}
}
\value{
A labelled data frame of the same dimensions as ESSPU containing
the calculated ESPPU weights.
}
\description{
This function converts ecosystem service provision potential per unit
scores to weights by dividing by a common denominator, or a matrix of
custom divisors per habitat/ecosystem service combination.
}
\section{Warning}{

If \code{custom_divisor_matrix} is provided, it must have the same
dimensions as \code{esppu}.
}

