% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ns_data.R
\name{get_ns_cirm_list}
\alias{get_ns_cirm_list}
\title{Extract List of Condition Indicator Relevance Matrices (CIRMs)}
\usage{
get_ns_cirm_list(
  path,
  sheet_list,
  matrix_range,
  ci_ids,
  all_service_labels,
  all_habitat_labels
)
}
\arguments{
\item{path}{Path to Excel source.}

\item{sheet_list}{Vector of sheet indices to iterate through.}

\item{matrix_range}{Excel range for the binary matrices.}

\item{ci_ids}{Vector of indicator IDs.}

\item{all_service_labels}{Flat vector of all service labels.}

\item{all_habitat_labels}{Flat vector of all habitat labels.}
}
\description{
Extract List of Condition Indicator Relevance Matrices (CIRMs)
}
\keyword{internal}
