% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ns_data.R
\name{get_ns_es_tree}
\alias{get_ns_es_tree}
\title{Extract and Clean Ecosystem Service Tree}
\usage{
get_ns_es_tree(path, sheet, est_range, es_code_ranges, es_name_ranges)
}
\arguments{
\item{path}{Path to Excel source.}

\item{sheet}{Sheet index.}

\item{est_range}{Range for ES categories.}

\item{es_code_ranges}{Vector of ranges for ES codes.}

\item{es_name_ranges}{Vector of ranges for ES names.}
}
\description{
Extract and Clean Ecosystem Service Tree
}
\keyword{internal}
