% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ns_data.R
\name{get_ns_habitat_tree}
\alias{get_ns_habitat_tree}
\title{Extract and Clean Habitat Classification Tree}
\usage{
get_ns_habitat_tree(path, sheet, bh_range, dh_ranges)
}
\arguments{
\item{path}{Path to Excel source.}

\item{sheet}{Sheet index.}

\item{bh_range}{Range for broad habitats.}

\item{dh_ranges}{Vector of ranges for detailed habitats.}
}
\description{
Extract and Clean Habitat Classification Tree
}
\keyword{internal}
