% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ns_data.R
\name{get_ns_importance_scores}
\alias{get_ns_importance_scores}
\title{Extract Ecosystem Service Importance Scores}
\usage{
get_ns_importance_scores(path, sheet, importance_ranges, es_tree)
}
\arguments{
\item{path}{Path to Excel source.}

\item{sheet}{Sheet index.}

\item{importance_ranges}{Named list of Excel ranges for each service type.}

\item{es_tree}{The cleaned ES label tree.}
}
\description{
Extract Ecosystem Service Importance Scores
}
\keyword{internal}
