% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise_crs.R
\name{harmonise_crs}
\alias{harmonise_crs}
\title{Title}
\usage{
harmonise_crs(input_object, aoi_object)
}
\arguments{
\item{input_object}{A shapefile containing habitats data}

\item{aoi_object}{A shapefile delineating the extent of the area of interest}
}
\value{
The input object with CRS matched to the AOI.
}
\description{
Title
}
\examples{

wrong_crs_input <- sf::st_read(
  system.file("extdata", "wrong_crs_data.shp", package = "openNCAI"),
    quiet = TRUE)
aoi <- sf::st_read(
  system.file("extdata", "test_aoi.shp", package = "openNCAI"),
    quiet = TRUE)

crs_harmonised_output <- harmonise_crs(wrong_crs_input, aoi)

sf::st_crs(wrong_crs_input) # The original (mismatched) CRS
sf::st_crs(aoi)        # The target AOI CRS
sf::st_crs(crs_harmonised_output) # The output CRS (should match the AOI)


}
