% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ns_testing_data.R
\name{import_ns_testing_data}
\alias{import_ns_testing_data}
\title{Import Reference Data for Replication Testing}
\usage{
import_ns_testing_data(path, habitats_label_tree, es_label_tree, year_list)
}
\arguments{
\item{path}{A string representing the file path to the NatureScot .xlsx source.}

\item{habitats_label_tree}{A named list of habitat labels used for matrix row naming.}

\item{es_label_tree}{A named list of ecosystem service labels used for matrix column naming.}
}
\value{
A nested list containing:
\itemize{
  \item \code{ref_es_potential_base}: Ecosystem Service Potential Base matrix (Sheet 6).
  \item \code{ref_wellbeing_potential_base}: Well-being Base matrix (Sheet 7).
  \item \code{ref_total_indicator_relevances}: Total Indicator Relevances matrix (Sheet 74).
  \item \code{ref_all_year_sheets}: A list of yearly natural capital asset matrices.
  \item \code{ref_index_breakdowns}: Data frames containing the final index and its
      breakdowns by service type and habitat.
}
}
\description{
Imports selected sheets and ranges from the NatureScot NCAI spreadsheet that
contain the intermediate and final results of their calculations. This data
is used as a "gold standard" to ensure the `openNCAI` package replicates the
methodology reliably.
}
\keyword{internal}
