% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_shape.R
\name{load_shape}
\alias{load_shape}
\title{Make a spatial object from a shapefile}
\usage{
load_shape(file_path)
}
\arguments{
\item{file_path}{A file path string.}
}
\value{
A spatial object.
}
\description{
Takes the file path of a shape file, reads in using sf, and returns a
spatial object.
}
\examples{
shapefile_path <- system.file(
"extdata","test_data.shp", package = "openNCAI")
load_shape(shapefile_path)

}
