% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ns_data.R
\name{make_custom_divisor_matrix}
\alias{make_custom_divisor_matrix}
\title{Generate Custom Divisor Matrix}
\usage{
make_custom_divisor_matrix(
  all_habitat_labels,
  all_es_labels,
  habitats_to_adjust,
  services_to_adjust,
  usual_divisor,
  custom_divisor
)
}
\arguments{
\item{all_habitat_labels}{Vector of habitat labels.}

\item{all_es_labels}{Vector of service labels.}

\item{habitats_to_adjust}{Vector of habitat shorthands (e.g., "b1").}

\item{services_to_adjust}{Vector of service keywords for partial matching.}

\item{usual_divisor}{Numeric default divisor (e.g., 5).}

\item{custom_divisor}{Numeric adjustment divisor (e.g., 1).}
}
\description{
Generate Custom Divisor Matrix
}
\keyword{internal}
