% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_between_importance_scores}
\alias{ns_between_importance_scores}
\title{Ecosystem Service Importance Scores (between-service-type)}
\format{
A named numeric vector where values range from 0 to 20,
representing the relative importance of each ecosystem service type.
}
\usage{
ns_between_importance_scores
}
\description{
A set of scores denoting the importance of each type
group of ecosystem services to Scotland.
}
\keyword{datasets}
