% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_ci_scores}
\alias{ns_ci_scores}
\title{Condition Indicator Score Matrix}
\format{
A data frame:
\itemize{
  \item Rows represent years in the time series.
  \item Columns represent condition indicators (e.g., "Adult red grouse density").
  \item Values are numeric condition scores.
}
}
\usage{
ns_ci_scores
}
\description{
A matrix containing the raw condition indicator scores over the
years of the time series.
}
\keyword{datasets}
