% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_cirms_list}
\alias{ns_cirms_list}
\title{Condition Indicator Relevance Matrix List}
\format{
A named list of data frames:
\itemize{
  \item Names are condition indicator IDs matching \code{ns_indicator_directory}.
  \item Each data frame has habitats as rows and ecosystem services as columns.
  \item Values are binary (0 or 1) indicating relevance.
}
}
\usage{
ns_cirms_list
}
\description{
A set of matrices recording whether or not each condition indicator is
relevant to each habitat/ecosystem service combination.
}
\keyword{datasets}
