% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_custom_divisor_matrix}
\alias{ns_custom_divisor_matrix}
\title{Nature Scot Custom Divisor Matrix}
\format{
A data frame:
\itemize{
  \item Rows represent habitats matching the habitats label tree.
  \item Columns represent ecosystem services matching the ecosystem services label tree.
  \item Values are numeric divisors used for weight normalization.
}
}
\usage{
ns_custom_divisor_matrix
}
\description{
A matrix containing numbers by which the Ecosystem Service Potential per
Unit Scores will be divided to calculate a weight.
}
\keyword{datasets}
