% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_dirty_ci_names}
\alias{ns_dirty_ci_names}
\title{Dirty (Print) Condition Indicatory Names}
\format{
A character vector where items are the original number and name of
NatureScot's 38 condition indicators:
}
\usage{
ns_dirty_ci_names
}
\description{
Print version (no string cleaning) list of condition indicator names which
is used internally.
}
\keyword{datasets}
