% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_display_ci_names}
\alias{ns_display_ci_names}
\title{Display names of Condition Indicators}
\format{
A character vector where items are the original number and name of
NatureScot's 38 condition indicators:
}
\usage{
ns_display_ci_names
}
\description{
Display name (no string cleaning) list of condition indicator names which
is used internally.
}
\keyword{datasets}
