% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_display_es_label_tree}
\alias{ns_display_es_label_tree}
\title{Display name version of Ecosystem Services Label Tree}
\format{
A named list of character vectors:
\itemize{
  \item Names represent ecosystem service types (e.g., "Provisioning").
  \item Values are character vectors of specific ecosystem services (e.g., "Cultivated crops").
}
}
\usage{
ns_display_es_label_tree
}
\description{
Display name version (no string cleaning) version of \code{ns_es_label_tree}
which is used internally.
}
\keyword{datasets}
