% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_display_habitats_label_tree}
\alias{ns_display_habitats_label_tree}
\title{Display name version of Habitats Label Tree}
\format{
A named list of character vectors:
\itemize{
  \item Names represent broad habitat groups (e.g., "Coastal habitats").
  \item Values are character vectors of specific habitat types (e.g., "Coastal shingle").
}
}
\usage{
ns_display_habitats_label_tree
}
\description{
Display name version (no string cleaning) version of \code{ns_habitats_label_tree}
which is used internally.
}
\keyword{datasets}
