% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_es_label_tree}
\alias{ns_es_label_tree}
\title{Ecosystem Services Label Tree}
\format{
A named list of character vectors:
\itemize{
  \item Names represent ecosystem service types (e.g., "Provisioning").
  \item Values are character vectors of specific ecosystem services (e.g., "Cultivated crops").
}
}
\usage{
ns_es_label_tree
}
\description{
A hierarchical tree mapping ecosystem service types to specific ecosystem
services.
}
\keyword{datasets}
