% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_habitat_extent}
\alias{ns_habitat_extent}
\title{Habitat Extent for Scotland}
\format{
A data frame:
\itemize{
  \item Rows represent broad habitats matching the habitats label tree.
  \item Columns represent years (e.g., 2000 to 2022).
  \item Values are numeric representing area in hectares.
}
}
\usage{
ns_habitat_extent
}
\description{
A table containing the annual extent (area) of various habitats.
}
\keyword{datasets}
