% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_habitats_label_tree}
\alias{ns_habitats_label_tree}
\title{Habitats Label Tree}
\format{
A named list of character vectors:
\itemize{
  \item Names represent broad habitat groups (e.g., "Coastal habitats").
  \item Values are character vectors of specific habitat types (e.g., "Coastal shingle").
}
}
\usage{
ns_habitats_label_tree
}
\description{
A hierarchical tree mapping broad habitat groups to specific habitat types.
}
\keyword{datasets}
