% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_indicator_directory}
\alias{ns_indicator_directory}
\title{Indicator Directory}
\format{
A data frame with the following columns:
\itemize{
  \item \code{ci_id}: Unique condition indicator identifier.
  \item \code{provisioning}: Salience for Provisioning services.
  \item \code{regulation_and_maintenance}: Salience for Regulation and Maintenance services.
  \item \code{cultural}: Salience for Cultural services.
}
}
\usage{
ns_indicator_directory
}
\description{
A table recording the salience of habitat condition scores as
indicators of the likely flow of services under each ecosystem service type.
}
\keyword{datasets}
