% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_provision_per_unit_scores}
\alias{ns_provision_per_unit_scores}
\title{Provision Per Unit Scores}
\format{
A data frame:
\itemize{
  \item Rows represent habitats matching the habitats label tree.
  \item Columns represent ecosystem services matching the ecosystem services label tree.
  \item Values are numeric scores from 0 (No potential) to 5 (Maximum potential).
}
}
\usage{
ns_provision_per_unit_scores
}
\description{
A matrix of scores denoting the exemplary capacity of each habitat to provide
each ecosystem service.
}
\keyword{datasets}
