% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_within_importance_scores}
\alias{ns_within_importance_scores}
\title{Ecosystem Service Importance Scores (within-service-type)}
\format{
A named list of numeric vectors:
\itemize{
  \item Names represent ecosystem service types.
  \item Values are numeric scores (0-20) for specific services within that group.
}
}
\usage{
ns_within_importance_scores
}
\description{
Sets of scores denoting the importance of each ecosystem service
to Scotland, within its type group.
}
\keyword{datasets}
