% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_year_list}
\alias{ns_year_list}
\title{Year List}
\format{
A character vector of year labels (e.g., "2000", "2001").
}
\usage{
ns_year_list
}
\description{
A list of years for which data is available and over which the index will
be calculated.
}
\keyword{datasets}
