% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openNCAI-package.R
\docType{package}
\name{openNCAI-package}
\alias{openNCAI}
\alias{openNCAI-package}
\title{openNCAI: Calculates a Natural Capital Assets Index}
\description{
Calculates a regional natural capital assets index (NCAI) following the methodology designed by NatureScot for Scotland as described in Albon, Balana, Brooker & Eastwood (2014) \url{https://www.nature.scot/sites/default/files/2025-06/naturescot-commissioned-report-751.pdf} and McKenna et al. (2019) \doi{10.1016/J.ECOLIND.2019.105645}. Processes habitat extent and condition data alongside metadata and weighting systems to produce a yearly single figure indexed relative to a base-year value of 100.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oharakate/openNCAI}
  \item Report bugs at \url{https://github.com/oharakate/openNCAI/issues}
}

}
\author{
\strong{Maintainer}: Kate O'Hara \email{k.a.ohara@stir.ac.uk}

Authors:
\itemize{
  \item Chris Littleboy \email{chris.littleboy@stir.ac.uk}
}

}
\keyword{internal}
