% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_and_read_data_template.R
\name{prepare_template_matrix}
\alias{prepare_template_matrix}
\title{Prepare a blank data frame based on Habitat and ES trees}
\usage{
prepare_template_matrix(hab_tree, es_tree)
}
\description{
Prepare a blank data frame based on Habitat and ES trees
}
\keyword{internal}
