% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ns_testing_data.R
\name{read_ns_year_sheet}
\alias{read_ns_year_sheet}
\title{Read and Label a Specific Yearly Asset Sheet}
\usage{
read_ns_year_sheet(sheet, path, es_label_tree, habitats_label_tree)
}
\arguments{
\item{sheet}{The index or name of the Excel sheet to read.}

\item{path}{String file path to the .xlsx source.}

\item{es_label_tree}{Named list of ES labels.}

\item{habitats_label_tree}{Named list of habitat labels.}
}
\value{
A labeled data frame representing the asset matrix for a single year.
}
\description{
Read and Label a Specific Yearly Asset Sheet
}
\keyword{internal}
