% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ns_data.R
\name{read_the_ci_scores}
\alias{read_the_ci_scores}
\title{Read Yearly Condition Indicator Scores}
\usage{
read_the_ci_scores(path, sheet_list, vector_range, ci_ids)
}
\arguments{
\item{path}{Path to Excel source.}

\item{sheet_list}{Vector of sheet indices.}

\item{vector_range}{Range for the single-column score vectors.}

\item{ci_ids}{Vector of indicator IDs.}
}
\description{
Read Yearly Condition Indicator Scores
}
\keyword{internal}
