% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_and_read_data_template.R
\name{write_condition_scores_sheet}
\alias{write_condition_scores_sheet}
\title{Write Condition Scores Time-Series Sheet}
\usage{
write_condition_scores_sheet(
  wb,
  sheet_name,
  ci_names,
  years,
  source_data = NULL
)
}
\description{
Write Condition Scores Time-Series Sheet
}
\keyword{internal}
