% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_and_read_data_template.R
\name{write_extent_sheet}
\alias{write_extent_sheet}
\title{Write Habitat Extent Time-Series Sheet}
\usage{
write_extent_sheet(
  wb,
  sheet_name,
  hab_tree,
  years,
  hab_palette,
  source_data = NULL
)
}
\description{
Write Habitat Extent Time-Series Sheet
}
\keyword{internal}
