% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_and_read_data_template.R
\name{write_importance_sheet_with_data}
\alias{write_importance_sheet_with_data}
\title{Write the specialized Importance Weights sheet}
\usage{
write_importance_sheet_with_data(
  wb,
  sheet_name,
  es_tree,
  between_data,
  within_list,
  instruction_style,
  entry_style
)
}
\description{
Write the specialized Importance Weights sheet
}
\keyword{internal}
