% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_and_read_data_template.R
\name{write_indicator_directory}
\alias{write_indicator_directory}
\title{Write Indicator Directory Sheet}
\usage{
write_indicator_directory(
  wb,
  sheet_name,
  ci_names,
  es_types,
  source_data = NULL
)
}
\description{
Write Indicator Directory Sheet
}
\keyword{internal}
