% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_and_read_data_template.R
\name{write_input_matrix}
\alias{write_input_matrix}
\title{Write standard Matrix sheets (e.g., ES Potential or CI Relevance)}
\usage{
write_input_matrix(
  wb,
  sheet_name,
  data_df,
  hab_tree,
  es_tree,
  style_obj,
  hab_palette,
  thick_border_style,
  instruction_style,
  instruction
)
}
\description{
Write standard Matrix sheets (e.g., ES Potential or CI Relevance)
}
\keyword{internal}
