% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_and_read_data_template.R
\name{write_instructions_sheet}
\alias{write_instructions_sheet}
\title{Write a clean Instructions sheet at the start of the workbook}
\usage{
write_instructions_sheet(wb, sheet_name)
}
\description{
Write a clean Instructions sheet at the start of the workbook
}
\keyword{internal}
