% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncai.R
\name{calc_ncai_by_st}
\alias{calc_ncai_by_st}
\title{Calculate NCAI Broken Down by Ecosystem Service Type}
\usage{
calc_ncai_by_st(total_assets_matrix_list, es_label_tree, year_one = NULL, ...)
}
\arguments{
\item{total_assets_matrix_list}{A named list of annual asset data frames.}

\item{es_label_tree}{A named list where each element is a character
vector of ecosystem service labels (column names).}

\item{year_one}{Optional: the year to index around. Default is the first
year of the \code{year_list}.}

\item{...}{Additional arguments passed to \code{calc_ncai} (e.g., \code{smoothing_weights}).}
}
\value{
A list of NCAI data frames, one for each ecosystem service group.
}
\description{
Calculate the index subsetted by ecosystem service type.
}
\keyword{internal}
