% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_and_read_data_template.R
\name{create_ncai_template}
\alias{create_ncai_template}
\title{Create an NCAI Data Entry Template}
\usage{
create_ncai_template(
  template_out,
  habitats_label_tree,
  es_label_tree,
  ci_names,
  year_list,
  habitat_extent = NULL,
  esppu_scores = NULL,
  between_importance_scores = NULL,
  within_importance_scores = NULL,
  ci_scores = NULL,
  indicator_directory = NULL,
  ci_relevance_matrices = NULL
)
}
\arguments{
\item{template_out}{A string representing the file path where the .xlsx will be saved.}

\item{habitats_label_tree}{A named list of character vectors representing the habitat hierarchy. Names: broad habitats, typically EUNIS level 1. Character vector items: typically EUNIS level 2 habitats.}

\item{es_label_tree}{A named list of character vectors representing the ecosystem service hierarchy. Names: typically SEEA ecosystem service types. Character vector items: typically CICES-type ecosystem services.}

\item{ci_names}{A character vector of condition indicator names.}

\item{year_list}{A vector of years (numeric or character) to be included in the account.}

\item{habitat_extent}{Optional data frame of existing habitat extent data. Row and column order MUST matched label trees/year list.}

\item{esppu_scores}{Optional data frame of existing ES Potential Per Unit scores. Row and column order MUST match label trees/year list/condition indicator names.}

\item{between_importance_scores}{Optional list of broad ES importance weights. Row and column order MUST matched label trees.}

\item{within_importance_scores}{Optional list of weights for specific services.}

\item{ci_scores}{Optional data frame of existing condition indicator scores.}

\item{indicator_directory}{Optional data frame of existing indicator salience data.}

\item{ci_relevance_matrices}{Optional list of existing binary relevance matrices.}
}
\value{
Generates an Excel file at the specified path and returns a message of success.
}
\description{
Generates a highly formatted, protected Excel (.xlsx) workbook containing
the necessary sheets for a user to populate an openNCAI account. The template
includes internal logic for habitat and ecosystem service trees, automated
styling, and cell locking to ensure data integrity for subsequent re-import.
Note that the habitats and ecosystem service label trees passed here may be
used to subsequently read in the data with read_ncai_template().
Optional arguments to pre-populate the template with data assume that the
passed-in data matches those of the label trees, year list and condition
indicator list as relevant.
}
\examples{
\dontrun{
# 1. Define the habitat hierarchy
habitat_label_tree <- list(
  'B. Coastal Habitats' = c(
    "Coastal vegetated shingle",
    "Coastal dunes and sandy shores"
  ),
  'E. Grasslands' = c(
    "Dry Grasslands",
    "Mesic Grasslands"
  )
)

# 2. Define the ecosystem service hierarchy
ecosystem_service_label_tree <- list(
  'PROVISIONING' = c(
    "1.1 Cultivated Crops",
    "1.2 Reared Animals And Their Outputs"
  ),
  'CULTURAL' = c(
    "3.5. Existence & bequest"
  )
)

# 3. Define indicators and temporal scope
condition_indicator_list <- c("National Water Quality Index", "AgriSCOR")
list_of_years <- 2020:2025

# 4. Generate the template
# The resulting Excel file will have locked headers and editable data areas.
create_ncai_template(
  template_out = "NCAI_Entry_Template.xlsx",
  habitats_label_tree = habitat_label_tree,
  es_label_tree = ecosystem_service_label_tree,
  ci_names = condition_indicator_list,
  year_list = list_of_years
)
}
}
