% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yearly_potentials.R
\name{get_yearly_potential_wellbeing}
\alias{get_yearly_potential_wellbeing}
\title{Calculate Yearly Potential Wellbeing Contribution}
\usage{
get_yearly_potential_wellbeing(
  habitat_extent,
  year_one,
  wellbeing_potential_base,
  as_matrices = FALSE
)
}
\arguments{
\item{habitat_extent}{A matrix or data frame where columns represent years
and rows represent habitats.}

\item{year_one}{Character or Numeric. The baseline year for indexing.}

\item{wellbeing_potential_base}{A matrix or data frame of Wellbeing Potential Base values.}

\item{as_matrices}{Logical. If \code{TRUE}, returns annual matrices;
if \code{FALSE}, returns a smoothed index data frame.}
}
\value{
Depending on the value of \code{as_matrices}:
\itemize{
  \item If \code{FALSE}: A data frame with columns \code{raw_total},
    \code{raw_index}, and \code{smoothed_index}.
  \item If \code{TRUE}: A named list of annual wellbeing matrices.
}
}
\description{
Calculates the potential wellbeing contribution time series by multiplying
indexed habitat extent by the Wellbeing Base. It can be understood as the
the potential wellbeing contribution of the habitat assets over the years,
before weighting by likely flow of ecosystem services based on habitat
condition.
}
\keyword{internal}
