% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ns_data.R
\name{import_ns_data}
\alias{import_ns_data}
\title{Import and Process Natural Capital Account Data}
\usage{
import_ns_data(path, year_list = 2000:2022, tir_constant = 2)
}
\arguments{
\item{path}{A string representing the file path to the .xlsx data source.}

\item{year_list}{A list of years, default = 2000:2022, numeric or character.}

\item{tir_constant}{Numeric constant to avoid zero divisions. Default is 2.}
}
\description{
Import and Process Natural Capital Account Data
}
