% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_ncai_matrix.R
\name{label_ncai_matrix}
\alias{label_ncai_matrix}
\title{Label NCAI Habitat/Ecosystem Service Matrix
This function takes a habitat/ecosystem service matrix of values used in
openNCAI and returns it as a data frame with the habitat and ecosystem
service labels applied.}
\usage{
label_ncai_matrix(matrix, habitats_label_tree, es_label_tree)
}
\arguments{
\item{matrix}{A matrix of values, e.g. scores, weights, where rows are
habitats and columns are ecosystem services}

\item{habitats_label_tree}{A named list of character vectors representing the
hierarchy of habitats (as character vectors, typically
EUNIS Level 2) within broad habitats (as list object names, typically EUNIS
Level 1). Syntactical names only (no spaces or special characters).}

\item{es_label_tree}{A named list of character vectors representing
the hierarchy of ecosystem services (as character vectors) within
service type group (as list object names).
Syntactical names only (no spaces or special characters).}
}
\value{
A labelled data frame.
}
\description{
Label NCAI Habitat/Ecosystem Service Matrix
This function takes a habitat/ecosystem service matrix of values used in
openNCAI and returns it as a data frame with the habitat and ecosystem
service labels applied.
}
\keyword{internal}
