% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openNCAI-package.R
\docType{package}
\name{openNCAI-package}
\alias{openNCAI}
\alias{openNCAI-package}
\title{openNCAI: Calculates a Natural Capital Assets Index}
\description{
openNCAI is an R package for calculating a regional natural capital assets index (NCAI), using the method designed by NatureScot to calculate Scotland's NCAI. It uses habitat extent and condition data, along with metadata and systems of weights, and produces a yearly single figure indexed around a year one value of 100.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oharakate/openNCAI}
  \item Report bugs at \url{https://github.com/oharakate/openNCAI/issues}
}

}
\author{
\strong{Maintainer}: Kate O'Hara \email{k.a.ohara@stir.ac.uk}

}
\keyword{internal}
