% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openNCAI-package.R
\docType{package}
\name{openNCAI-package}
\alias{openNCAI}
\alias{openNCAI-package}
\title{openNCAI: Calculates a Natural Capital Assets Indicator for a particular area of interest}
\description{
Calculates natural capital assets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oharakate/openNCAI}
  \item Report bugs at \url{https://github.com/oharakate/openNCAI/issues}
}

}
\keyword{internal}
