% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openNCAI-package.R
\docType{package}
\name{openNCAI-package}
\alias{openNCAI}
\alias{openNCAI-package}
\title{openNCAI: Calculates a Natural Capital Assets Index}
\description{
Calculates a regional natural capital assets index (NCAI) following the methodology designed by 'NatureScot' for 'Scotland'. Processes habitat extent and condition data alongside metadata and weighting systems to produce a yearly single figure indexed relative to a base-year value of 100.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oharakate/openNCAI}
  \item Report bugs at \url{https://github.com/oharakate/openNCAI/issues}
}

}
\author{
\strong{Maintainer}: Kate O'Hara \email{k.a.ohara@stir.ac.uk}

}
\keyword{internal}
